requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

# Runtime (server/API) requirements
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
tqdm>=4.65.0

# HTTP / Web
requests>=2.28.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
jinja2>=3.1.0

# Optional: include for CLI/testing convenience
# starlette is pulled in by fastapi; add extras here if needed
